# Conflicts:
#	src/main/java/com/tcdt/qlnvhang/service/xuathang/kiemtrachatluong/XhHoSoKyThuatService.java

+++ src/main/java/com/tcdt/qlnvhang/entities/xuathang/daugia/xuatkho/XhDgBangKeDtl.java
@@ -5,6 +5,7 @@ import lombok.Data;
 
 import javax.persistence.*;
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name = XhDgBangKeDtl.TABLE_NAME)
@@ -21,6 +22,6 @@ public class XhDgBangKeDtl extends BaseEntity implements Serializable {
   private Long idHdr;
   private String maCan;
   private String soBaoBi;
-  private Long trongLuongBaoBi;
-  private Long trongLuongCaBi;
+  private BigDecimal trongLuongBaoBi;
+  private BigDecimal trongLuongCaBi;
 }
